refactor: extract shared card utils, tighten github validation, add docs
- Extract getMonogram, getMonogramBg, getHueRotate, getAvatarUrl,
  getAvatarFromSource into shared $lib/card-utils.ts module
- Remove duplicate implementations from McpServersTab and SkillCard
- Tighten github catalogId regex to match GitHub naming rules
- Add clarifying comments for deduplication set and media query scope
